services/chat-service/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
drf-yasg
python-dotenv
psycopg2-binary
django-redis
channels
channels-redis
redis
daphne
gunicorn
confluent-kafka
requests

